Write a program that multiplies two numbers

+++ 0x0A-argc_argv/main.h
@@ -7,6 +7,6 @@ int _putchar(char str);
 void print_program_name(char *program_name);
 void print_program_number(int num);
 void print_args(int argc, char *argv[]);
-unsigned int mul(size_t n, size_t m);
+int mul(int n, int m);
 
 #endif
